Desktop/hybrid/requirements.txt
# Autochess Hybrid - Python Dependencies

# UI / Rendering
pygame-ce>=2.5.0

# Testing Framework
pytest>=9.0.0
pytest-cov>=4.0.0

# Code Quality
pylint>=3.0.0
black>=24.0.0
flake8>=7.0.0

# Utilities
colorama>=0.4.6

# Type Checking (optional)
mypy>=1.8.0
